Move Map.TilesByDistance to MapGrid and make non-static.
